+++ src/containers/navigation-icons/NavigationIcons.js
@@ -2,8 +2,8 @@ import { useContext, useCallback } from 'react'
 import { useSelector } from 'react-redux'
 
 import Loader from '~/components/loader/Loader'
-import GuestIcons from '~/containers/navigation-icons/GuestIcons'
-import StudentIcons from '~/containers/navigation-icons/StudentIcons'
+import GuestIcons from '~/containers/navigation-icons/guest-icons/GuestIcons'
+import StudentIcons from '~/containers/navigation-icons/student-icons/StudentIcons'
 import { ModalContext } from '~/context/modal-context'
 import LoginDialog from '~/containers/guest-home-page/login-dialog/LoginDialog'
 
@@ -19,7 +19,7 @@ const NavigationIcons = ({ setIsSidebarOpen }) => {
 
   if (loading) return(<Loader size={ 20 } />)
 
-  if (userRole === 'student') return(<StudentIcons setIsSidebarOpen={ setIsSidebarOpen } />)
+  if (userRole === 'student') return (<StudentIcons setIsSidebarOpen={ setIsSidebarOpen } />)
 
   return (<GuestIcons openLoginDialog={ openLoginDialog }  setIsSidebarOpen={ setIsSidebarOpen } />)
