+++ src/containers/navigation-icons/guest-icons/GuestIcons.js
